+++ backend/IntexBackendApi/Models/DonationAllocation.cs
@@ -11,7 +11,7 @@ public class DonationAllocation
     public int DonationId { get; set; }
     public int SafehouseId { get; set; }
     public string? ProgramArea { get; set; }
-    public decimal? AmountAllocated { get; set; }
+    public double? AmountAllocated { get; set; }
     public DateOnly? AllocationDate { get; set; }
     public string? AllocationNotes { get; set; }
 }
